[ADD] 플레이어 속도 증가, 룸 관련 오류 해결
현재 문을 막는 코드에서 ground타일맵이 안 없어지는 현상 발견

+++ Assets/Scripts/Monster/MonsterMovement.cs
@@ -12,7 +12,7 @@ public class MonsterMovement : MonoBehaviour
     private MonsterManager monsterManager;
     private MonsterData monsterData;
 
-    private int currentHealth; // 현재 체력
+    private float currentHealth; // 현재 체력
 
     private void Start()
     {
@@ -79,7 +79,7 @@ public class MonsterMovement : MonoBehaviour
     }
 
     // 대미지를 받는 함수
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
         Debug.Log(monsterData.name + "이(가) " + damage + " 대미지를 받았습니다. 현재 체력: " + currentHealth);
